fix crash without grs
